refactor: add recursive rank task with optional rating seed
